add documentation and structure to hierarchy tutorials Signed-off-by: dbulahov <bulahovdenis@gmail.com>
